Extract Messages class
