+++ cucumber-messages/docgen/README.md
@@ -1,0 +1,4 @@
+# Documentation Generation for Cucumber Protocol messages
+
+This directory contains scripts for generating documentation and test messages
+from `.proto` files.
