+++ cucumber-messages/generate-examples.rb
@@ -10,47 +10,10 @@
 # about the messages. It's generated with pseudomuto/protoc-gen-doc
 require 'json'
 require 'fileutils'
+require_relative 'docgen/message'
 
-messages = {}
-meta = JSON.parse(IO.read(File.dirname(__FILE__) + '/messages.json'))
-meta['files'].each do |file|
-  file['messages'].each do |message|
-    long_name = message['longName']
-    messages[long_name] = message
-  end
-end
-
-class Message
-  def initialize(long_name, messages)
-    @long_name = long_name
-    @messages = messages
-  end
-  
-  def example
-    m = @messages[@long_name]
-    raise "No such message: #{@long_name}" unless m
-    e = {}
-    m['fields'].each do |field|
-      field_name = field['name']
-      long_type = field['longType']
-      repeated = field['label'] == 'repeated'
-      field_examples = []
-      if @messages[long_type]
-        field = Message.new(long_type, @messages)
-        field_examples.push(field.example)
-      else
-        description = field['description']
-        description.split(/\n/).each do |line|
-          if /^\s*Example:\s*(.*)/ =~ line
-            field_examples.push(JSON.parse($1))
-          end
-        end
-      end
-      e[field_name] = repeated ? field_examples : field_examples[0]
-    end
-    e
-  end
-end
+messages_hash = JSON.parse(IO.read(File.dirname(__FILE__) + '/messages.json'))
+messages = Message.all(messages_hash)
 
 FileUtils.rm_rf("examples")
 FileUtils.mkdir("examples")
